feat: extract prestashop version service from update config
